requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
pydantic==2.6.0
pydantic-settings==2.1.0
faiss-cpu==1.9.0.post1
scikit-learn==1.3.2
numpy==1.26.3
rapidfuzz==3.6.1
sentence-transformers==2.6.1
python-dotenv==1.0.1
transformers==4.40.2
torch==2.0.1
requests==2.32.5
tqdm==4.67.3

